Added confirm to product delete and fixed plan page header on small screens - 338, 164

+++ src/components/Products.vue
@@ -357,16 +357,18 @@ export default {
      * Deletes the selected products.
      */
     async deleteProducts () {
-      try {
-        this.$store.commit('setData', {
-          attr: 'dontLeave',
-          data: true
-        })
-        await this.$store.dispatch('deleteProduct', this.selectedProducts)
-        this.deselectAll()
-        this.$store.dispatch('endLoading')
-      } catch (e) {
-        this.$parent.$parent.resolveError(e)
+      if (await this.$parent.$parent.$parent.$refs.confirm_pop_up.show('Are you sure that you want to delete all the selected products?', 'We will remove these products from our database and it won\'t be recoverable.')) {
+        try {
+          this.$store.commit('setData', {
+            attr: 'dontLeave',
+            data: true
+          })
+          await this.$store.dispatch('deleteProduct', this.selectedProducts)
+          this.deselectAll()
+          this.$store.dispatch('endLoading')
+        } catch (e) {
+          this.$parent.$parent.resolveError(e)
+        }
       }
     },
